update load way

+++ lib/has_magic_fields/extend.rb
@@ -111,14 +111,14 @@ module HasMagicFields
       end
     end
 
-
-    %w{ models }.each do |dir|
-      path = File.join(File.dirname(__FILE__), '../app', dir)
-      $LOAD_PATH << path
-      ActiveSupport::Dependencies.autoload_paths << path
-      ActiveSupport::Dependencies.autoload_once_paths.delete(path)
+    if Rails.env.development?
+      %w{ models }.each do |dir|
+        path = File.join(File.dirname(__FILE__), '../app', dir)
+        $LOAD_PATH << path
+        ActiveSupport::Dependencies.autoload_paths << path
+        ActiveSupport::Dependencies.autoload_once_paths.delete(path)
+      end
     end
-    
   end
 end
